requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

setuptools~=60.2.0
tensorflow~=2.12.0
keras~=2.12.0
numpy~=1.23.5
Pillow~=9.5.0